requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Limiter==3.7.0
PyMySQL==1.1.0
cryptography==41.0.4
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==11.0.0
psycopg2-binary==2.9.9; python_version < "3.13"
gunicorn>=20.1.0
